# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MathMind AI EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /practice/questions:
    get:
      operationId: listQuestions
      tags: [practice]
      summary: Get practice questions
      parameters:
        - name: topicId
          in: query
          schema:
            type: integer
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /practice/questions/{id}/submit:
    post:
      operationId: submitAnswer
      tags: [practice]
      summary: Submit answer for a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswerBody"
      responses:
        "200":
          description: Answer result with AI explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerResult"
  /practice/generate:
    post:
      operationId: generateQuestion
      tags: [practice]
      summary: AI-generate a personalized question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuestionBody"
      responses:
        "200":
          description: Generated question with step-by-step solution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedQuestion"

  # Progress
...
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all tutor conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new tutor conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a message to the AI tutor (streaming)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
components:
  schemas:
...
    Question:
      type: object
      properties:
        id:
          type: integer
        topicId:
          type: integer
        text:
          type: string
        textHi:
          type: ["string", "null"]
        options:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        points:
          type: integer
        timeLimit:
          type: integer
      required:
        - id
        - topicId
        - text
        - options
        - difficulty
        - points
        - timeLimit
    SubmitAnswerBody:
      type: object
      properties:
        answer:
          type: string
        timeTaken:
          type: integer
      required:
        - answer
        - timeTaken
    AnswerResult:
      type: object
      properties:
        correct:
          type: boolean
        correctAnswer:
          type: string
        xpEarned:
          type: integer
        explanation:
          type: string
        explanationHi:
          type: ["string", "null"]
        steps:
          type: array
          items:
            type: string
      required:
        - correct
        - correctAnswer
        - xpEarned
        - explanation
        - steps
    GenerateQuestionBody:
      type: object
      properties:
        topicId:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        language:
          type: string
          enum: [en, hi]
      required:
        - topicId
        - difficulty
    GeneratedQuestion:
      type: object
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        steps:
          type: array
          items:
            type: string
      required:
        - question
        - options
        - correctAnswer
        - explanation
        - steps

    # Progress
...
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
        language:
          type: string
          enum: [en, hi]
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error